REFACTOR: Change the path of a dataclass
Changed the path of the dataclass "CategoryNameWithId" from
stylish/domain/model/viewModel to stylish/domain/model.
