+++ app/src/main/java/com/example/stylish/domain/viewModel/ShardViewModel.kt
@@ -6,6 +6,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.stylish.data.remote.dto.category.Category
 import com.example.stylish.data.remote.dto.category.Children
+import com.example.stylish.domain.model.CategoryNameWithId
 import com.example.stylish.domain.repository.AsosRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
